requirements.txt
black==21.12b0
click==8.0.3
Flask==2.0.2
Flask-Moment==1.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.0.0
platformdirs==2.4.1
python-dotenv==0.19.2
SQLAlchemy==1.4.31
tomli==1.2.3
typing-extensions==4.0.1
Werkzeug==2.0.2
WTForms==3.0.1
